change jan-util.js to jan-util.ts
