+++ src/main/ts/jan/mcr-util.js
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 import HandUtil from './hand-util';
-import JanUtil from './jan-util';
+import {hasJanpai} from './jan-util';
 import {McrComplete} from './mcr-complete';
 import {MCR_YAKU} from './mcr-yaku';
 import {sortMentsuList, toJanpaiList} from './mentsu-util';
@@ -455,7 +455,7 @@ export default class McrUtil {
     
     static _removeWaitYaku(yakuList, knittedChowList, janpai) {
         if (yakuList.includes(MCR_YAKU.ALL_CHOWS) && yakuList.includes(MCR_YAKU.KNITTED_STRAIGHT) &&
-            JanUtil.hasJanpai(toJanpaiList(knittedChowList), janpai)) {
+            hasJanpai(toJanpaiList(knittedChowList), janpai)) {
             this._removeYaku(yakuList, MCR_YAKU.EDGE_WAIT);
             this._removeYaku(yakuList, MCR_YAKU.CLOSED_WAIT);
             this._removeYaku(yakuList, MCR_YAKU.SINGLE_WAIT);
